perf(log-tracker): lazy tabs loading for better loading times
