+++ apps/client/src/app/pages/log-tracker/log-tracker/log-tracker.component.less
@@ -24,6 +24,12 @@
   .ant-divider-horizontal {
     margin: 8px 0;
   }
+
+  .ant-tabs-tab-btn {
+    a {
+      color: var(--text-color) !important;
+    }
+  }
 }
 
 .dol-path-buttons {
